A package is renamed from power_spectral_density to spectral_processing in order to be consistent with the specifications. All classes are updated if needed.
