+++ src/ansys/sound/core/spectral_processing/__init__.py
@@ -25,7 +25,7 @@
 Helper functions related to Power Spectral Density (PSD) calculations.
 """
 
-from ._power_spectral_density_parent import PowerSpectralDensityParent
+from ._spectral_processing_parent import SpectralProcessingParent
 from .power_spectral_density import PowerSpectralDensity
 
-__all__ = ("PowerSpectralDensityParent", "PowerSpectralDensity")
+__all__ = ("SpectralProcessingParent", "PowerSpectralDensity")
