+++ tests/tests_power_spectral_density/test_power_spectral_density.py
@@ -27,7 +27,7 @@ import numpy as np
 import pytest
 
 from ansys.sound.core._pyansys_sound import PyAnsysSoundException, PyAnsysSoundWarning
-from ansys.sound.core.power_spectral_density import PowerSpectralDensity
+from ansys.sound.core.spectral_processing import PowerSpectralDensity
 from ansys.sound.core.signal_utilities import LoadWav
